Revert "fuchsia: Add scaffolding and basic implementation for flatland migration (#27423)"
This reverts commit d5b410ec502a97ed7ceff05da63a0e97e9daee06.
