+++ shell/platform/fuchsia/flutter/BUILD.gn
@@ -59,12 +59,6 @@ template("runner_sources") {
       "component.h",
       "engine.cc",
       "engine.h",
-      "flatland_connection.cc",
-      "flatland_connection.h",
-      "flatland_external_view_embedder.cc",
-      "flatland_external_view_embedder.h",
-      "flatland_platform_view.cc",
-      "flatland_platform_view.h",
       "flutter_runner_product_configuration.cc",
       "flutter_runner_product_configuration.h",
       "focus_delegate.cc",
@@ -73,8 +67,6 @@ template("runner_sources") {
       "fuchsia_external_view_embedder.h",
       "fuchsia_intl.cc",
       "fuchsia_intl.h",
-      "gfx_platform_view.cc",
-      "gfx_platform_view.h",
       "gfx_session_connection.cc",
       "gfx_session_connection.h",
       "isolate_configurator.cc",
@@ -137,7 +129,6 @@ template("runner_sources") {
              "$fuchsia_sdk_root/fidl:fuchsia.io",
              "$fuchsia_sdk_root/fidl:fuchsia.sys",
              "$fuchsia_sdk_root/fidl:fuchsia.ui.app",
-             "$fuchsia_sdk_root/fidl:fuchsia.ui.composition",
              "$fuchsia_sdk_root/fidl:fuchsia.ui.scenic",
              "$fuchsia_sdk_root/pkg:async-cpp",
              "$fuchsia_sdk_root/pkg:async-default",
