+++ shell/platform/fuchsia/flutter/engine.h
@@ -9,7 +9,6 @@
 
 #include <fuchsia/intl/cpp/fidl.h>
 #include <fuchsia/io/cpp/fidl.h>
-#include <fuchsia/ui/composition/cpp/fidl.h>
 #include <fuchsia/ui/gfx/cpp/fidl.h>
 #include <fuchsia/ui/views/cpp/fidl.h>
 #include <lib/async-loop/cpp/loop.h>
@@ -25,8 +24,6 @@
 #include "flutter/shell/common/thread_host.h"
 #include "flutter/shell/platform/fuchsia/flutter/accessibility_bridge.h"
 
-#include "flatland_connection.h"
-#include "flatland_external_view_embedder.h"
 #include "flutter_runner_product_configuration.h"
 #include "fuchsia_external_view_embedder.h"
 #include "gfx_session_connection.h"
@@ -50,7 +47,6 @@ class Engine final {
 
   static flutter::ThreadHost CreateThreadHost(const std::string& name_prefix);
 
-  // Gfx connection ctor.
   Engine(Delegate& delegate,
          std::string thread_label,
          std::shared_ptr<sys::ServiceDirectory> svc,
@@ -61,19 +57,6 @@ class Engine final {
          UniqueFDIONS fdio_ns,
          fidl::InterfaceRequest<fuchsia::io::Directory> directory_request,
          FlutterRunnerProductConfiguration product_config);
-
-  // Flatland connection ctor.
-  Engine(Delegate& delegate,
-         std::string thread_label,
-         std::shared_ptr<sys::ServiceDirectory> svc,
-         std::shared_ptr<sys::ServiceDirectory> runner_services,
-         flutter::Settings settings,
-         fuchsia::ui::views::ViewCreationToken view_creation_token,
-         scenic::ViewRefPair view_ref_pair,
-         UniqueFDIONS fdio_ns,
-         fidl::InterfaceRequest<fuchsia::io::Directory> directory_request,
-         FlutterRunnerProductConfiguration product_config);
-
   ~Engine();
 
   // Returns the Dart return code for the root isolate if one is present. This
@@ -90,19 +73,9 @@ class Engine final {
   const std::string thread_label_;
   flutter::ThreadHost thread_host_;
 
-  std::optional<VulkanSurfaceProducer> surface_producer_;
-
-  // Gfx specific classes.
-  fuchsia::ui::views::ViewToken view_token_;
   std::shared_ptr<GfxSessionConnection> session_connection_;
+  std::optional<VulkanSurfaceProducer> surface_producer_;
   std::shared_ptr<FuchsiaExternalViewEmbedder> external_view_embedder_;
-
-  // Flatland specific classes.
-  fuchsia::ui::views::ViewCreationToken view_creation_token_;
-  std::shared_ptr<FlatlandConnection> flatland_connection_;
-  std::shared_ptr<FlatlandExternalViewEmbedder> flatland_view_embedder_;
-
-  scenic::ViewRefPair view_ref_pair_;
 
   std::unique_ptr<IsolateConfigurator> isolate_configurator_;
   std::unique_ptr<flutter::Shell> shell_;
@@ -115,15 +88,6 @@ class Engine final {
   bool intercept_all_input_ = false;
 
   fml::WeakPtrFactory<Engine> weak_factory_;
-
-  void Initialize(
-      bool use_flatland,
-      std::shared_ptr<sys::ServiceDirectory> svc,
-      std::shared_ptr<sys::ServiceDirectory> runner_services,
-      flutter::Settings settings,
-      UniqueFDIONS fdio_ns,
-      fidl::InterfaceRequest<fuchsia::io::Directory> directory_request,
-      FlutterRunnerProductConfiguration product_config);
 
   static void WarmupSkps(
       fml::BasicTaskRunner* concurrent_task_runner,
