+++ shell/platform/fuchsia/flutter/meta/flutter_aot_runner.cmx
@@ -24,8 +24,6 @@
             "fuchsia.sysmem.Allocator",
             "fuchsia.timezone.Timezone",
             "fuchsia.tracing.provider.Registry",
-            "fuchsia.ui.composition.Allocator",
-            "fuchsia.ui.composition.Flatland",
             "fuchsia.vulkan.loader.Loader"
         ]
     }
