+++ shell/platform/fuchsia/flutter/meta/flutter_jit_product_runner.cmx
@@ -25,8 +25,6 @@
             "fuchsia.sysmem.Allocator",
             "fuchsia.timezone.Timezone",
             "fuchsia.tracing.provider.Registry",
-            "fuchsia.ui.composition.Allocator",
-            "fuchsia.ui.composition.Flatland",
             "fuchsia.vulkan.loader.Loader"
         ]
     }
