+++ shell/platform/fuchsia/flutter/vulkan_surface.cc
@@ -103,17 +103,16 @@ bool VulkanSurface::CreateVulkanImage(vulkan::VulkanProvider& vulkan_provider,
 VulkanSurface::VulkanSurface(
     vulkan::VulkanProvider& vulkan_provider,
     fuchsia::sysmem::AllocatorSyncPtr& sysmem_allocator,
-    fuchsia::ui::composition::AllocatorPtr& flatland_allocator,
     sk_sp<GrDirectContext> context,
     scenic::Session* session,
     const SkISize& size,
     uint32_t buffer_id)
     : vulkan_provider_(vulkan_provider), session_(session), wait_(this) {
-  FML_CHECK(session_ || flatland_allocator.is_bound());
+  FML_CHECK(session_ != nullptr);
   FML_CHECK(context != nullptr);
 
-  if (!AllocateDeviceMemory(sysmem_allocator, flatland_allocator,
-                            std::move(context), size, buffer_id)) {
+  if (!AllocateDeviceMemory(sysmem_allocator, std::move(context), size,
+                            buffer_id)) {
     FML_LOG(ERROR) << "VulkanSurface: Could not allocate device memory.";
     return;
   }
@@ -135,15 +134,11 @@ VulkanSurface::VulkanSurface(
 }
 
 VulkanSurface::~VulkanSurface() {
-  if (session_) {
-    if (image_id_) {
-      session_->Enqueue(scenic::NewReleaseResourceCmd(image_id_));
-    }
-    if (buffer_id_) {
-      session_->DeregisterBufferCollection(buffer_id_);
-    }
-  } else {
-    release_image_callback_();
+  if (image_id_) {
+    session_->Enqueue(scenic::NewReleaseResourceCmd(image_id_));
+  }
+  if (buffer_id_) {
+    session_->DeregisterBufferCollection(buffer_id_);
   }
   wait_.Cancel();
   wait_.set_object(ZX_HANDLE_INVALID);
@@ -232,7 +227,6 @@ bool VulkanSurface::CreateFences() {
 
 bool VulkanSurface::AllocateDeviceMemory(
     fuchsia::sysmem::AllocatorSyncPtr& sysmem_allocator,
-    fuchsia::ui::composition::AllocatorPtr& flatland_allocator,
     sk_sp<GrDirectContext> context,
     const SkISize& size,
     uint32_t buffer_id) {
@@ -256,29 +250,8 @@ bool VulkanSurface::AllocateDeviceMemory(
   LOG_AND_RETURN(status != ZX_OK,
                  "VulkanSurface: Failed to sync collection token");
 
-  if (session_) {
-    session_->RegisterBufferCollection(buffer_id, std::move(scenic_token));
-    buffer_id_ = buffer_id;
-  } else {
-    fuchsia::ui::composition::BufferCollectionExportToken export_token;
-    status =
-        zx::eventpair::create(0, &export_token.value, &import_token_.value);
-
-    fuchsia::ui::composition::RegisterBufferCollectionArgs args;
-    args.set_export_token(std::move(export_token));
-    args.set_buffer_collection_token(std::move(scenic_token));
-    args.set_usage(
-        fuchsia::ui::composition::RegisterBufferCollectionUsage::DEFAULT);
-    flatland_allocator->RegisterBufferCollection(
-        std::move(args),
-        [](fuchsia::ui::composition::Allocator_RegisterBufferCollection_Result
-               result) {
-          if (result.is_err()) {
-            FML_LOG(ERROR)
-                << "RegisterBufferCollection call to Scenic Allocator failed";
-          }
-        });
-  }
+  session_->RegisterBufferCollection(buffer_id, std::move(scenic_token));
+  buffer_id_ = buffer_id;
 
   VkBufferCollectionCreateInfoFUCHSIA import_info;
   import_info.collectionToken = vulkan_token.Unbind().TakeChannel().release();
@@ -406,18 +379,10 @@ bool VulkanSurface::SetupSkiaSurface(sk_sp<GrDirectContext> context,
 }
 
 void VulkanSurface::PushSessionImageSetupOps(scenic::Session* session) {
-  if (session) {
-    if (image_id_ == 0)
-      image_id_ = session->AllocResourceId();
-    session->Enqueue(scenic::NewCreateImage2Cmd(
-        image_id_, sk_surface_->width(), sk_surface_->height(), buffer_id_, 0));
-  }
-}
-
-void VulkanSurface::SetImageId(uint32_t image_id) {
-  FML_CHECK(image_id_ == 0);
-  FML_CHECK(!session_);
-  image_id_ = image_id;
+  if (image_id_ == 0)
+    image_id_ = session->AllocResourceId();
+  session->Enqueue(scenic::NewCreateImage2Cmd(
+      image_id_, sk_surface_->width(), sk_surface_->height(), buffer_id_, 0));
 }
 
 uint32_t VulkanSurface::GetImageId() {
@@ -428,33 +393,6 @@ sk_sp<SkSurface> VulkanSurface::GetSkiaSurface() const {
   return valid_ ? sk_surface_ : nullptr;
 }
 
-fuchsia::ui::composition::BufferCollectionImportToken
-VulkanSurface::GetBufferCollectionImportToken() {
-  FML_CHECK(!session_);
-  fuchsia::ui::composition::BufferCollectionImportToken import_dup;
-  import_token_.value.duplicate(ZX_RIGHT_SAME_RIGHTS, &import_dup.value);
-  return import_dup;
-}
-
-zx::event VulkanSurface::GetAcquireFence() {
-  FML_CHECK(!session_);
-  zx::event fence;
-  acquire_event_.duplicate(ZX_RIGHT_SAME_RIGHTS, &fence);
-  return fence;
-}
-
-zx::event VulkanSurface::GetReleaseFence() {
-  FML_CHECK(!session_);
-  zx::event fence;
-  release_event_.duplicate(ZX_RIGHT_SAME_RIGHTS, &fence);
-  return fence;
-}
-void VulkanSurface::SetReleaseImageCallback(
-    ReleaseImageCallback release_image_callback) {
-  FML_CHECK(!session_);
-  release_image_callback_ = release_image_callback;
-}
-
 size_t VulkanSurface::AdvanceAndGetAge() {
   size_history_[size_history_index_] = GetSize();
   size_history_index_ = (size_history_index_ + 1) % kSizeHistorySize;
@@ -463,16 +401,15 @@ size_t VulkanSurface::AdvanceAndGetAge() {
 }
 
 bool VulkanSurface::FlushSessionAcquireAndReleaseEvents() {
-  if (session_) {
-    zx::event acquire, release;
-    if (acquire_event_.duplicate(ZX_RIGHT_SAME_RIGHTS, &acquire) != ZX_OK ||
-        release_event_.duplicate(ZX_RIGHT_SAME_RIGHTS, &release) != ZX_OK) {
-      return false;
-    }
-    session_->EnqueueAcquireFence(std::move(acquire));
-    session_->EnqueueReleaseFence(std::move(release));
+  zx::event acquire, release;
+
+  if (acquire_event_.duplicate(ZX_RIGHT_SAME_RIGHTS, &acquire) != ZX_OK ||
+      release_event_.duplicate(ZX_RIGHT_SAME_RIGHTS, &release) != ZX_OK) {
+    return false;
   }
 
+  session_->EnqueueAcquireFence(std::move(acquire));
+  session_->EnqueueReleaseFence(std::move(release));
   age_ = 0;
   return true;
 }
